+++ eventregistry/__init__.py
@@ -1,1 +1,2 @@
 ﻿from eventregistry.EventRegistry import *
+from eventregistry._version import __version__